#3 [fix] 병합문제 해결
